Fix bug where using a 16-bit signal mask caused some signal handlers not to be installed.
This manifested itself in interactive hush shells getting suspended when ^Z was typed (and maybe also other ways).

+++ shell/hush.c
@@ -8278,7 +8278,7 @@ static void install_special_sighandlers(void)
 	}
 	/* Careful, do not re-install handlers we already installed */
 	if (G.special_sig_mask != mask) {
-		unsigned diff = mask & ~G.special_sig_mask;
+		unsigned long diff = mask & ~G.special_sig_mask;
 		G.special_sig_mask = mask;
 		install_sighandlers(diff);
 	}
